// src/application/shared/cnpj/dto/consulta-cnpj-response.dto.ts
export class ConsultaCnpjResponseDto {
    cnpj: string;
    nome: string;
    fantasia: string;
    endereco: string;
    abertura: string;
    situacao: string;
    tipo: string;
    porte: string;
    natureza_juridica: string;
    logradouro: string;
    numero: string;
    municipio: string;
    bairro: string;
    uf: string;
    cep: string;
    email: string;
    telefone: string;
}